tests: Move test_auto_subbed_to_personals test to test_signup.py.
This is really a test about account creation.
